test updates - currently suffering security manager issues
"gradle -Dtest.single=*CoordinatedSyncConflictMDDeleteMMSUpdate test"

+++ src/main/java/gov/nasa/jpl/mbee/mdk/api/MagicDrawHelper.java
@@ -54,6 +54,9 @@ import com.nomagic.utils.ErrorHandler;
 import gov.nasa.jpl.mbee.mdk.ems.ReferenceException;
 import gov.nasa.jpl.mbee.mdk.lib.Utils;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -76,6 +79,45 @@ public class MagicDrawHelper {
         project = Application.getInstance().getProject();
         ef = project.getElementsFactory();
     }
+
+    /*****************************************************************************************
+     *
+     * Project load / close functions
+     *
+     *****************************************************************************************/
+
+    public static ProjectDescriptor openProject(String filename) throws IOException {
+        final File file = new File(filename);
+        if (!file.isFile()) {
+            throw new IOException(Paths.get(file.toURI()).toString() + " is not a file and cannot be loaded.");
+        }
+        final ProjectDescriptor projectDescriptor = ProjectDescriptorsFactory.createProjectDescriptor(file.toURI());
+        if (projectDescriptor == null) {
+            throw new IOException(Paths.get(file.toURI()).toString() + " could not generate a project descriptor.");
+        }
+        final ProjectsManager projectsManager = Application.getInstance().getProjectsManager();
+        projectsManager.loadProject(projectDescriptor, true);
+        final Project project = projectsManager.getActiveProject();
+        if (project == null) {
+            throw new IOException(Paths.get(file.toURI()).toString() + " could not be loaded into MagicDraw.");
+        }
+        return projectDescriptor;
+    }
+
+    public static void saveProject(String filename) {
+        System.out.println(System.currentTimeMillis());
+        File file = new File(filename);
+        ProjectDescriptor projectDescriptor =
+                ProjectDescriptorsFactory.createLocalProjectDescriptor(Application.getInstance().getProject(), file);
+        Application.getInstance().getProjectsManager().saveProject(projectDescriptor, false);
+        System.out.println(System.currentTimeMillis());
+    }
+
+    public static void closeProject() {
+        Application.getInstance().getProjectsManager().closeProject();
+    }
+
+
 
     /*****************************************************************************************
      *
@@ -208,17 +250,15 @@ public class MagicDrawHelper {
      */
     public static void deleteEditableContainerChildren(Element parent) throws ReadOnlyElementException {
         Collection<Element> elements = parent.getOwnedElement();
-        ArrayList<Element> eList = new ArrayList<Element>();
+        ArrayList<Element> eList = new ArrayList<>();
 
         for (Element e : elements) {
-            if (e.getHumanName().startsWith("Package") && e.isEditable()) {
-                eList.add(e);
-            }
-            if (e.getHumanName().startsWith("Diagram") && e.isEditable()) {
+            if (e.canBeDeleted() && e.isEditable() && (e.getHumanName().startsWith("Diagram")
+                    || (e.getHumanName().startsWith("Package")
+                    && !(e.getHumanName().contains("__MMSSync__") || e.getHumanName().contains("Holding Bin"))))) {
                 eList.add(e);
             }
         }
-
         for (Element elem : eList) {
             deleteMDElement(elem);
         }
@@ -227,7 +267,11 @@ public class MagicDrawHelper {
     /**
      * Deletes all editable container elements in a project.
      */
-    public static void deleteLocalMDElements() throws ReadOnlyElementException {
+    public static void clearModel() throws ReadOnlyElementException {
+        if (ElementFinder.getModelRoot() == null) {
+            System.out.println("Null root");
+            return;
+        }
         deleteEditableContainerChildren(ElementFinder.getModelRoot());
     }
